+++ backend/substrapp/tests/compute_tasks/test_compute_task.py
@@ -1,4 +1,4 @@
-import itertools
+import datetime
 import uuid
 from typing import Optional
 from typing import Type
@@ -34,22 +34,6 @@ def test_is_task_status_runnable_allow_doing():
     assert task_utils._is_task_status_runnable(compute_task, allow_doing=True)
 
 
-RUNNABLE_COMPUTE_PLAN_STATUSES = task_utils._RUNNABLE_COMPUTE_PLAN_STATUSES
-COMPUTE_PLAN_STATUSES = [status for status in orchestrator.ComputePlanStatus]
-NON_RUNNABLE_COMPUTE_PLAN_STATUSES = [
-    status for status in COMPUTE_PLAN_STATUSES if status not in RUNNABLE_COMPUTE_PLAN_STATUSES
-]
-
-
-@pytest.mark.parametrize(
-    ("compute_plan_status", "is_runnable"),
-    [(s, True) for s in RUNNABLE_COMPUTE_PLAN_STATUSES] + [(s, False) for s in NON_RUNNABLE_COMPUTE_PLAN_STATUSES],
-)
-def test_is_compute_plan_status_runnable(compute_plan_status: int, is_runnable: bool):
-    compute_plan = orc_mock.ComputePlanFactory(status=compute_plan_status)
-    assert task_utils._is_compute_plan_status_runnable(compute_plan) is is_runnable
-
-
 @pytest.fixture
 def client() -> mock.Mock:
     return mock.Mock(spec=orc_client.OrchestratorClient)
@@ -64,38 +48,46 @@ def test_raise_if_task_not_runnable_raise_TaskNonRunnableStatusError(task_status
         assert exc.status == task_status
 
 
-@pytest.mark.parametrize("compute_plan_status", NON_RUNNABLE_COMPUTE_PLAN_STATUSES)
-def test_raise_if_task_not_runnable_raise_ComputePlanNonRunnableStatusError(  # noqa: N802
-    compute_plan_status: str, client: mock.Mock
+@pytest.mark.parametrize(
+    ("compute_plan_cancelation_date", "compute_plan_failure_date"),
+    [
+        (datetime.datetime.now(), None),
+        (None, datetime.datetime.now()),
+    ],
+)
+def test_raise_if_task_not_runnable_raise_ComputePlanNonRunnableError(  # noqa: N802
+    compute_plan_cancelation_date: Optional[datetime.datetime],
+    compute_plan_failure_date: Optional[datetime.datetime],
+    client: mock.Mock,
 ):
     task_status = RUNNABLE_TASK_STATUSES[0]
     task = orc_mock.ComputeTaskFactory(status=task_status, compute_plan_key="cp-key")
-    client.query_compute_plan.return_value = orc_mock.ComputePlanFactory(status=compute_plan_status)
+    client.query_compute_plan.return_value = orc_mock.ComputePlanFactory(
+        cancelation_date=compute_plan_cancelation_date,
+        failure_date=compute_plan_failure_date,
+    )
 
-    with pytest.raises(task_utils.ComputePlanNonRunnableStatusError) as exc:
+    with pytest.raises(task_utils.ComputePlanNonRunnableError) as exc:
         task_utils._raise_if_task_not_runnable(str(uuid.uuid4()), client, existing_task=task)
-        assert exc.status == compute_plan_status
+        assert exc.cancelation_date == compute_plan_cancelation_date
+        assert exc.failure_date == compute_plan_failure_date
 
     client.query_compute_plan.assert_called_once()
 
 
-@pytest.mark.parametrize(
-    ("task_status", "compute_plan_status"),
-    list(itertools.product(RUNNABLE_TASK_STATUSES, RUNNABLE_COMPUTE_PLAN_STATUSES)),
-)
-def test_raise_if_task_not_runnable_do_not_raise(task_status: str, compute_plan_status: str, client: mock.Mock):
+@pytest.mark.parametrize("task_status", RUNNABLE_TASK_STATUSES)
+def test_raise_if_task_not_runnable_do_not_raise(task_status: str, client: mock.Mock):
     task = orc_mock.ComputeTaskFactory(status=task_status, compute_plan_key="cp-key")
-    client.query_compute_plan.return_value = orc_mock.ComputePlanFactory(status=compute_plan_status)
+    client.query_compute_plan.return_value = orc_mock.ComputePlanFactory()
 
     task_utils._raise_if_task_not_runnable(str(uuid.uuid4()), client, existing_task=task)
 
     client.query_compute_plan.assert_called_once()
 
 
-@pytest.mark.parametrize("compute_plan_status", RUNNABLE_COMPUTE_PLAN_STATUSES)
-def test_raise_if_task_not_runnable_allow_doing_do_not_raise(compute_plan_status: str, client: mock.Mock):
+def test_raise_if_task_not_runnable_allow_doing_do_not_raise(client: mock.Mock):
     task = orc_mock.ComputeTaskFactory(status=orchestrator.ComputeTaskStatus.STATUS_DOING, compute_plan_key="cp-key")
-    client.query_compute_plan.return_value = orc_mock.ComputePlanFactory(status=compute_plan_status)
+    client.query_compute_plan.return_value = orc_mock.ComputePlanFactory()
 
     task_utils._raise_if_task_not_runnable(str(uuid.uuid4()), client, allow_doing=True, existing_task=task)
 
@@ -103,7 +95,7 @@ def test_raise_if_task_not_runnable_allow_doing_do_not_raise(compute_plan_status
 
 
 @pytest.mark.parametrize(
-    "exception", [task_utils.TaskNonRunnableStatusError, task_utils.ComputePlanNonRunnableStatusError, None]
+    "exception", [task_utils.TaskNonRunnableStatusError, task_utils.ComputePlanNonRunnableError, None]
 )
 @mock.patch("substrapp.compute_tasks.compute_task.raise_if_task_not_runnable", autospec=True)
 def is_task_runnable(_raise_if_task_not_runnable: mock.Mock, exception: Optional[Type[Exception]]):