+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Don't use the `status` field of the compute plan protocol buffer messages.
 - BREAKING: remove `delete_intermediary_models` field from the `compute_plan` view.
 - Use task output asset in API response
 - Add channel column to input/output tables.